@@ -1,55 +1,51 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using Cronos;
 
-namespace CronBackgroundServices
+namespace CronBackgroundServices;
+
+internal class Timing
 {
-    internal class Timing
+    public readonly TimeZoneInfo TimeZoneInfo;
+
+    public Timing(string timeZoneId)
     {
-        public readonly TimeZoneInfo TimeZoneInfo;
+        TimeZoneInfo = TimeZoneInfo.FindSystemTimeZoneById(timeZoneId);
+    }
 
-        public Timing(string timeZoneId)
-        {
-            TimeZoneInfo = TimeZoneInfo.FindSystemTimeZoneById(timeZoneId);
-        }
+    public DateTimeOffset RelativeNow(DateTimeOffset? nowutc = null)
+    {
+        return TimeZoneInfo.ConvertTime(nowutc ?? DateTimeOffset.UtcNow, TimeZoneInfo);
+    }
 
-        public DateTimeOffset RelativeNow(DateTimeOffset? nowutc = null)
-        {
-            return TimeZoneInfo.ConvertTime(nowutc ?? DateTimeOffset.UtcNow, TimeZoneInfo);
-        }
+    public DateTimeOffset? GetNextOccurenceInRelativeTime(string cron)
+    {
+        var expression = CronExpression.Parse(cron, CronFormat.IncludeSeconds);
+        return expression.GetNextOccurrence(DateTimeOffset.UtcNow, TimeZoneInfo);
+    }
 
-        public DateTimeOffset? GetNextOccurenceInRelativeTime(string cron)
-        {
-            var expression = CronExpression.Parse(cron, CronFormat.IncludeSeconds);
-            return expression.GetNextOccurrence(DateTimeOffset.UtcNow, TimeZoneInfo);
-        }
+    public IEnumerable<string> Get10NextOccurrences(string cron)
+    {
+        var expression = CronExpression.Parse(cron, CronFormat.IncludeSeconds);
+        var fromUtc = DateTime.UtcNow;
+        var upcoming = new List<DateTime>();
+        upcoming.AddRange(Get10Occurrences(upcoming, expression, fromUtc, fromUtc.AddMonths(1)));
+        return upcoming.Select(u => $"{u.ToLongDateString()} {u.ToLongTimeString()}");
+    }
 
-        public IEnumerable<string> Get10NextOccurrences(string cron)
+    private IEnumerable<DateTime> Get10Occurrences(List<DateTime> upcoming, CronExpression expression, DateTime fromUtc, DateTime toUtc)
+    {
+        while (true)
         {
-            var expression = CronExpression.Parse(cron, CronFormat.IncludeSeconds);
-            var fromUtc = DateTime.UtcNow;
-            var upcoming = new List<DateTime>();
-            upcoming.AddRange(Get10Occurrences(upcoming, expression, fromUtc, fromUtc.AddMonths(1)));
-            return upcoming.Select(u => $"{u.ToLongDateString()} {u.ToLongTimeString()}");
-        }
+            toUtc = toUtc.AddMonths(1);
+            var occurrences = expression.GetOccurrences(fromUtc, toUtc);
+            upcoming = occurrences.ToList();
 
-        private IEnumerable<DateTime> Get10Occurrences(List<DateTime> upcoming, CronExpression expression, DateTime fromUtc, DateTime toUtc)
-        {
-            while (true)
+            if (upcoming.Count < 10)
             {
-                toUtc = toUtc.AddMonths(1);
-                var occurrences = expression.GetOccurrences(fromUtc, toUtc);
-                upcoming = occurrences.ToList();
-
-                if (upcoming.Count < 10)
-                {
-                    continue;
-                }
-                break;
+                continue;
             }
-            return upcoming.Take(10);
-
+            break;
         }
+        return upcoming.Take(10);
+
     }
 }
